Simplify method names
The changes help distinguish between command and event handler methods by using the "for" and "on" prefixes respectively. Furthermore, the return type is removed from the method signature as it is not needed given that the compiler will check that the correct type is returned.

A small internal benefit is that the code size of the macro is reduced slightly given the commonality of method names without consideration of their return type.