Permits the conveying of multiple states given an event change
During the crafting of a new FSM in my app, I noticed that we hadn't catered for supporting the scenario where a state transition given an event could yield multiple states. This scenario also manifests itself in another app where we aren't using the macro here.

To achieve this, I introduce a means of expressing the multitude of types that an event transition can produce. While I could have used the more conventional inferred type ("_"), I feel it is important to retain the documentation here. One of the goals of this project is to have the macros convey a diagram as closely as it can. An inferred type would prevent a diagram from being generated.

The error reporting is nice. If anything other than a State type is returned from the event handler then the compiler tells you. Similarly, if you provide a single type then the compiler (via the macro) will ensure that it is an enum variant (this was there before).