Entry/exit called only if they are discriminant
Prior to this commit, entry/exit handlers were called even when transitioning to the same state. The change here is to discriminant, accepting that if the state S is not an enum then the behaviour is undefined (the macros actually require S to be enum, so this should be fine).