Mutable state and associated types (#12)
* Mutable state and associated types

State is now passed in as being mutable so that some handlers may efficiently update state. A consequence of mutable state is that we can no longer have exit handlers, otherwise we would have to clone state so that we can pass it to an exit handler once event processing has occurred. On having analysed our own usage of edfsm, we felt that exit handlers can generally be replaced with entry handlers, which are retained. There is a tradeoff when it comes to having removed exit handlers. It can mean that more entry handlers are required, but we feel that being able to mutate state in some circumstances presents a much larger benefit.

Mutating state also permits us to clearly identify a transition vs an update to state. Consequently, the `is_transitioning` function is no longer required. This simplifies our code base further.

Associated types are now used in place of generics. One reason for this is so that a state machine can be confined to representing just one FSM. Associated types also make it easier to pass FSMs around and compose them with other structures as they are self-contained.