Allow side effects to be async
I've often found that side effects of the FSM are required to be async. A scenario that has come up regularly has been to send some command on a channel from a side effect. Prior to the changes here, that send would have to be a "try send" and possibly fail if the channel buffer is full. With the async changes here, we can back pressure the send until the buffer frees up.

By providing an async command handler, the application developer can chose whether a side effect is to be async or not. The overhead from being non-async to async should be negligible. In the case where sync behaviour is still required, the time take in returning a future that is immediately ready should be very small.

A consequence of the change is that all of the FSM types (state, command...) are required to implement Send.